backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.50.0
langchain>=0.3.0
langchain-core>=1.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
youtube-transcript-api>=0.6.0
sentence-transformers>=2.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
requests>=2.28.0
